requirements.txt
paho-mqtt
requests
pyyaml
flask
schedule
soco
